+++ common/src/main/java/tocraft/walkers/traits/impl/CantSwimTrait.java
@@ -1,0 +1,28 @@
+package tocraft.walkers.traits.impl;
+
+import com.mojang.serialization.MapCodec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.LivingEntity;
+import tocraft.walkers.Walkers;
+import tocraft.walkers.traits.ShapeTrait;
+
+public class CantSwimTrait<E extends LivingEntity> extends ShapeTrait<E> {
+    public static final ResourceLocation ID = Walkers.id("cant_swim");
+    public static final MapCodec<CantSwimTrait<?>> CODEC = RecordCodecBuilder.mapCodec((instance) -> instance.stable(new CantSwimTrait<>()));
+
+    @Override
+    public ResourceLocation getId() {
+        return ID;
+    }
+
+    @Override
+    public MapCodec<? extends ShapeTrait<?>> codec() {
+        return CODEC;
+    }
+
+    @Override
+    public boolean canBeRegisteredMultipleTimes() {
+        return false;
+    }
+}